adicionanado a lista de games dinamica

+++ app/src/main/java/com/example/testfragment/model/Game.kt
@@ -1,14 +1,13 @@
 package com.example.testfragment.model
 
 import android.graphics.drawable.Drawable
-import java.io.FileDescriptor
 
 data class Game(
     var id:Long,
     var gameName: String,
     var gameImage: Drawable,
     var gameDescription: String,
-    var gameRating: Double,
+    var gameRating: Float,
     var console: List<Console>
 ) {
